raw.csv file upload.
This file consists of interesting data that the Fraper algorithm pulled directly from Facebook which was launched on to the HitchHiqe network as of November 23, 2019. 
To be noted : raw.csv won't be updated again. 
To be noted: The last column, Type, says if a post is one from a driver or rider. This was done manually. 0 -> Driver; 1 -> Rider; -1  -> UNDEFINED